Move GetProjectionAsync
