+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/RazorDocumentMappingService.cs
@@ -5,6 +5,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Razor.Language;
+using Microsoft.AspNetCore.Razor.LanguageServer.Extensions;
 using Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
 
@@ -29,5 +30,36 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer
         public abstract RazorLanguageKind GetLanguageKind(RazorCodeDocument codeDocument, int originalIndex, bool rightAssociative);
 
         public abstract Task<WorkspaceEdit> RemapWorkspaceEditAsync(WorkspaceEdit workspaceEdit, CancellationToken cancellationToken);
+
+        public async Task<Projection> GetProjectionAsync(DocumentContext documentContext, int absoluteIndex, CancellationToken cancellationToken)
+        {
+            var codeDocument = await documentContext.GetCodeDocumentAsync(cancellationToken).ConfigureAwait(false);
+            var sourceText = await documentContext.GetSourceTextAsync(cancellationToken).ConfigureAwait(false);
+
+            sourceText.GetLineAndOffset(absoluteIndex, out var line, out var character);
+            var projectedPosition = new Position(line, character);
+
+            var languageKind = GetLanguageKind(codeDocument, absoluteIndex, rightAssociative: false);
+            if (languageKind == RazorLanguageKind.CSharp)
+            {
+                if (TryMapToProjectedDocumentPosition(codeDocument, absoluteIndex, out var mappedPosition, out _))
+                {
+                    // For C# locations, we attempt to return the corresponding position
+                    // within the projected document
+                    projectedPosition = mappedPosition;
+                }
+                else
+                {
+                    // It no longer makes sense to think of this location as C#, since it doesn't
+                    // correspond to any position in the projected document. This should not happen
+                    // since there should be source mappings for all the C# spans.
+                    languageKind = RazorLanguageKind.Razor;
+                }
+            }
+
+            return new Projection(languageKind, projectedPosition, absoluteIndex);
+        }
     }
+
+    internal record Projection(RazorLanguageKind LanguageKind, Position Position, int AbsoluteIndex);
 }
